requirements.txt
google-generativeai
pydub
python-dotenv
fastapi
uvicorn
python-multipart
